Fine_tuned_approach/requirements.txt
lightrag-hku
neo4j
sentence-transformers
openai
python-dotenv
pytest
pypdf
python-docx
faiss-cpu
rich
flask[async]
flask-cors
fastapi
uvicorn
python-multipart
asyncpg
psycopg2-binary
pgvector